fix: merging add-new-post
