--- src/features/createPost/index.tsx
@@ -1,1 +1,0 @@
-export { PostCreationWizard } from './modal/PostCreationWizard';
