--- src/features/createPost/modal/AddPostSteps/CropStep/CropStep.tsx
@@ -1,133 +1,0 @@
-'use client';
-
-import Cropper from 'react-easy-crop';
-import { Modal } from '@/widgets/modal/Modal';
-import { Button, IconSprite } from '@/shared/ui';
-import { RootState } from '@/shared/state/store';
-import { useSelector } from 'react-redux';
-import { CroppingModalHeader } from './CroppingModalHeader';
-import { CroppingModalSettings } from './CroppingModalSettings';
-import { cn } from '@/shared/lib/cn';
-import { aspectRatios } from '@/shared/config/aspectRatios';
-import { ThumbnailsPreview } from './ThumbnailsPreview';
-import { usePhotoNavigation } from '@/features/createPost/hooks/usePhotoNavigation';
-import { useImageCropping } from '@/features/createPost/hooks/useImageCropping';
-import { useImageEditor } from '@/features/createPost/hooks/useImageEditor';
-import { useImageSave } from '@/features/createPost/hooks/useImageSave';
-import { useImageAspect } from '@/features/createPost/hooks/useImageAspect';
-import { PhotoNavigation } from '@/features/createPost/ui/PhotoNavigation';
-
-type CropStep = { onClose: () => void };
-
-export function CropStep({ onClose }: CropStep): React.ReactElement {
-  const photos = useSelector((state: RootState) => state.post.photos);
-  const currentIndex = useSelector(
-    (state: RootState) => state.post.currentPhotoIndex
-  );
-  const currentPhoto = photos[currentIndex];
-  const { hasNext, hasPrev, goNext, goPrev } = usePhotoNavigation();
-  const { onCropComplete, handleCropChange, croppedAreaPixels } =
-    useImageCropping(currentPhoto);
-  const {
-    croppedViewSettings,
-    setZoom,
-    setRotation,
-    handleToggleSetting,
-    handleCropRatioChange,
-  } = useImageEditor(currentPhoto);
-  const { handleNext } = useImageSave(currentPhoto);
-  const { getCropperAspect } = useImageAspect(currentPhoto);
-
-  return (
-    <Modal
-      onClose={onClose}
-      headerContent={
-        <CroppingModalHeader
-          stepToGo="filters"
-          stepToBack="upload"
-          onNext={() => handleNext(croppedAreaPixels)}
-        />
-      }
-      size="md"
-      open={true}
-      showCloseButton={false}
-    >
-      <div className="flex flex-col items-center">
-        <div className="relative h-[400px] w-full">
-          <Cropper
-            image={currentPhoto.url}
-            key={`cropper-${currentPhoto.url}-${currentIndex}`}
-            crop={currentPhoto.crop}
-            zoom={currentPhoto.zoom}
-            rotation={currentPhoto.rotation}
-            onCropChange={handleCropChange}
-            onZoomChange={setZoom}
-            onCropComplete={onCropComplete}
-            restrictPosition={true}
-            aspect={getCropperAspect()}
-            classes={{
-              containerClassName: 'bg-black rounded-lg',
-              mediaClassName: 'object-contain',
-              cropAreaClassName: 'border-2 border-blue-500',
-            }}
-          />
-          <div className="">
-            <PhotoNavigation
-              hasPrev={hasPrev}
-              hasNext={hasNext}
-              onPrev={goPrev}
-              onNext={goNext}
-            />
-          </div>
-        </div>
-
-        <div className="mt-4 flex w-full max-w-md justify-between">
-          <CroppingModalSettings
-            zoom={currentPhoto.zoom}
-            rotation={currentPhoto.rotation}
-            croppedViewSettings={croppedViewSettings}
-            aspectRatios={aspectRatios.map((ratio) =>
-              ratio.label === 'Original'
-                ? { ...ratio, value: currentPhoto.naturalAspect }
-                : ratio
-            )}
-            onZoomChange={setZoom}
-            onRotationChange={setRotation}
-            onToggleSetting={handleToggleSetting}
-            onCropRatioChange={handleCropRatioChange}
-            currentCropRatio={currentPhoto.cropRatio}
-          />
-
-          <div>
-            <div className="relative">
-              <Button
-                variant="text"
-                className={cn(
-                  'bg-dark-500 group w-0 border-0 focus:border-0 active:border-0'
-                )}
-                onClick={() => handleToggleSetting('thumbnails')}
-              >
-                <IconSprite
-                  iconName={'image-outline'}
-                  className={cn(
-                    'fill-light-100 transition-colors duration-200',
-                    {
-                      'fill-accent-500': croppedViewSettings.thumbnails,
-                    }
-                  )}
-                />
-              </Button>
-
-              {croppedViewSettings.thumbnails && (
-                <ThumbnailsPreview
-                  photos={photos.map((photo) => photo.url)}
-                  currentIndex={currentIndex}
-                />
-              )}
-            </div>
-          </div>
-        </div>
-      </div>
-    </Modal>
-  );
-}
