--- src/features/createPost/modal/ExitConfirmation.tsx
@@ -1,27 +1,0 @@
-import { ConfirmationDialog } from './ConfirmationDialog';
-import { JSX } from 'react';
-
-type ExitConfirmationProps = {
-  open: boolean;
-  onConfirm: () => void;
-  onCancel: () => void;
-};
-
-export function ExitConfirmation({
-  open,
-  onConfirm,
-  onCancel,
-}: ExitConfirmationProps): JSX.Element {
-  return (
-    <ConfirmationDialog
-      open={open}
-      title="Close"
-      message={`Do you really want to close the creation of a publication?
-If you close everything will be deleted`}
-      confirmText="Discard"
-      cancelText="Cancel"
-      onConfirm={onConfirm}
-      onCancel={onCancel}
-    />
-  );
-}
