+++ src/features/postCreate/index.tsx
@@ -1,0 +1,1 @@
+export { PostCreationWizard } from './ui/PostCreationWizard';
