+++ src/features/postCreate/ui/PhotoNavigation.tsx
